fixed error with type (task 2)

+++ Homework/task2/QuadraticEquation/Program.cs
@@ -1,8 +1,8 @@
 ﻿using System;
 
-double a = Convert.ToInt32(Console.ReadLine());
-double b = Convert.ToInt32(Console.ReadLine());
-double c = Convert.ToInt32(Console.ReadLine());
+double a = Convert.ToDouble(Console.ReadLine());
+double b = Convert.ToDouble(Console.ReadLine());
+double c = Convert.ToDouble(Console.ReadLine());
 
 Equations solver = new Equations(a, b, c);
